+++ R/NMextractDataFile.R
@@ -12,7 +12,7 @@
 ##'     and the existence of the file in that directory is not
 ##'     checked.
 ##' @param file.mod The input control stream. Default is to look for
-##'     \"file\" with extension changed to .mod (PSN style). You can
+##'     \"file\" with extension changed to `.mod` (PSN style). You can
 ##'     also supply the path to the file, or you can provide a
 ##'     function that translates the output file path to the input
 ##'     file path. The default behavior can be configured using