+++ NEWS.md
@@ -1,11 +1,65 @@
 
 # 0.1.7
+
+## New features
+* `NMreadPartab()` has been generalized to support comment formats very
+  generally. `NMreadPartab()` reads the comments in `$THETA`, `$OMEGA`
+  and `$SIGMA` sections, splits them into variables, and organizes
+  those variables in a parameter table. With this upgrade, pretty much
+  any structure should be supported as long as delimitors are not
+  alphabetic or numeric (so any special characters should
+  work). Notice, delimitors can change between fields . Example:
+  "$THETA 1.4 ; 3 - CL (Clearance) [L/h]" would be matched by
+  `NMreadPartab(...,format="%init ;%idx-%symbol(%label)[%unit]")`
+  which would then return a table including columns init, idx, symbol,
+  label, and unit. The comments must be systematic within say `$THETA`
+  but the format can be different for `$OMEGA` and `$SIGMA`. See
+  examples in `?NMreadParTab`.
+  
+* `NMrelate()` is a new automated approach to label parameters. It
+  interprets Nonmem code and provides labels used in the control
+  stream. If the line `TVCL=THETA(1)` is the only line in the code
+  that references THETA(1), `NMrelate()` will return a label
+  `TVCL`.
+
+* `mergeCheck()` has additional features available in the common.cols
+  argument.
+
+* `NMreadExt()` separates objective function values into a separate list
+  element. The `return` argument is used to control what data to
+  retrieve. Use one of "pars" (default, parameter estimates),
+  "iterations" (parameter estimates for each iteration), "obj" for
+  objective funtion value, or "all" for a list with all of those.
+  
+* `NMreadExt()` adds block information to `OMEGA` and `SIGMA` elements
+  based on off-diagonal values.
+  
+* `NMreadExt()` adds a `par.name` column which is provides consistent
+  parameter naming. Instead of Nonmem's `THETA1` which is found in the
+  `parameter` column, the `par.name` column will contain `THETA(1)`
+  consistent with the `OMEGA` and `SIGMA` naming like `OMEGA(1,1)`
+
+* `NMreadExt()` recognizes Laplacian estimation steps in addition to
+  the already supported FO, FOCE(i), SAEM, and IMP.
+
+* A new option `nc` can be controlled with NMdataConf(). This is to
+  serve `NMsim`. Please see `NMsim::NMexec`. `NMsim::NMsim()` does not
+  adhere to this setting because it does not parallellize by default.
+
 ## Bugfixes
 * `NMscanInput()` and `NMreadCsv()` could fail if file names had no
   extensions. Fixed.
 
 * `NMreplaceDataFile()` now works on directories and regular
   expressions to find models to update.
+  
+* Some internal functions would make some functions including
+  `NMscanData()` fail if used within `lapply()`. Fixed.
+
+* `NMexpandDoses()` would give a warning if `length(cols.id)>1`. Fixed.
+
+* `NMreadExt()` would mess up iterations and parameter estimates if
+  `as.fun` was set to returning something else than `data.table`s. Fixed.
 
 # 0.1.6